% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{compute_mixprobs1}
\alias{compute_mixprobs1}
\title{Title}
\usage{
compute_mixprobs1(Zs, unique_z, pi_h, Delta, mu, alpha, Lambda, Sigma, inds)
}
\arguments{
\item{inds}{}
}
\value{

}
\description{
Title
}
