% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Software.R
\name{GMC_Impute}
\alias{GMC_Impute}
\title{Fit Gaussian mixture copula with the margin adjustment and impute missing data}
\usage{
GMC_Impute(
  Data,
  nImp = 10,
  Impute = T,
  H = 25,
  k.star = NULL,
  nsamp = 10000,
  burn = floor(0.5 * nsamp),
  hyperparams = NULL,
  seed = NULL
)
}
\arguments{
\item{Data}{Data frame that is the input into the Gaussian Mixture Copula. Categorical and  binary variables should be factors}

\item{nImp}{Number of imputations to Create}

\item{Impute}{logical; Whether or not to create multiple imputations}

\item{H}{Upper bound for the number of components of the mixtures}

\item{k.star}{Dimension of latent factor in mixture}

\item{nsamp}{Number of samples for the MCMC}

\item{burn}{Number of burn-in samples before imputations}

\item{hyperparams}{Optional; Hyperparameter specifications for finite mixuture and RPL
 with default values given by:
\itemize{
\item \code{k.star}: dimension of latent factors, default is \code{floor(.7*p))}
\item \code{plugin.threshold}: threshold for number of unique values in each column for the re-sampling step to be conducted,
default is 350
\item \code{a_alpha}: prior for the stick breaking precision parameter, where the prior is given by \code{gamma(a_alpha,b_alpha)},
default value is 1
\item \code{b_alpha}: prior for the stick breaking precision parameter, where the prior is given by \code{gamma(a_alpha,b_alpha)},
default value is 3
\item \code{nu_mix}: hyper parameter for the component specific prior \eqn{N-IW(0,I_{k.star},nu_mix,kappa_0)} on \eqn{\eta},
default is k.star +2
\item \code{kappa_0}: hyper parameter for the component specific prior \eqn{N-IW(0,I_{k.star},nu_mix,kappa_0)} on \eqn{\eta},
default is .001
\item \code{nu}: hyper parameter for local shrinkage of MGP prior on components of Lambda, default is 3
\item \code{a1}: hyper parameter for global shrinkage of MGP prior on components of Lambda, default is 2
\item \code{a2}: hyper parameter for global shrinkage of MGP prior on components of Lambda, default is 3
\item \code{a.sigma,b.sigma}: hyper parameter for gamma prior on error variance, default value is 1 and 0.3
\item \code{delta}: prior precision parameter for covariance matrix in latent mixture. Increase to discover fewer clusters.
default value is 10
\item \code{D_0}: prior covariance for N-IW. Default is a k.star dimenional identity
}}
}
\value{
A list with the following elements:
\itemize{
\item \code{Imputations} nImp complete data sets
\item \code{Deltas}: List of length H, each containing an \code{(nsamp - burn x k.star x k.star)} array of
 posterior samples of cluster specific covariance matrices
\item \code{mus}: List of length H, each containing an \code{(nsamp-burn x k.star)} array of posterior samples of cluster
specific covariance matrices
\item \code{Sigmas}: array of dimension \code{(nsamp-burn x p)} containing posterior samples of the error covariance
\item \code{Quantiles}:List of length p, each containing an array of dimension \code{(nsamp-burn x length(support))} with posterior samples of point-wise cumulative probabilities.
\item \code{Support}: List of length p, each containing an array of dimension \code{(nsamp-burn x length(support))} with support of each variable
\item \code{col_mem}: array of length p containing variable names
\item \code{cat_col_names}: array containing names of categorical variables, with the number of repititions indicating eachs' number of levels
\item \code{bin_col_names}: array containing names of binary variables
\item \code{count_col_names}: array containing names of count-valued variables
\item \code{cont_col_names}: array containing names of continous variables
\item \code{dat}: Original data frame, useful for simulating predictive data sets
\item \code{seed}: Seed for replication of results
}
}
\description{
Fit the Gaussian mixture copula to a data set with or without missing data.
Return posterior samples of model parameters as well optional multiple imputations
}
\examples{
num = 500
X1<-rnorm(num)
X2<- rpois(num,5*abs(X1))
R = t(sapply(1:num, function(x)rbernoulli(2, p = pnorm(-.5 + .5*abs(X1[x]))))) #change to -.25 for more missing
X<- data.frame(X1,X2,X3)
X_noMis = X
X[which(R[,1] == T),2] = NA
X[which(R[,2] == T),3] = NA
imps<-GMC_Impute(Data = X, nsamp = 1000)
}
