% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Software.R
\name{get_predictive_Y}
\alias{get_predictive_Y}
\title{Henerate a Posterior Predictive Data Set}
\usage{
get_predictive_Y(
  n,
  Lambda,
  mu,
  alpha,
  Delta,
  pi_h,
  Sigma.diag,
  col_mem,
  cat_col_names,
  bin_col_names,
  count_col_names,
  cont_col_names,
  H,
  Y,
  z,
  Fns
)
}
\arguments{
\item{Fns}{}
}
\value{

}
\description{
Henerate a Posterior Predictive Data Set
}
