% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Software.R
\name{get_predictive_Y}
\alias{get_predictive_Y}
\title{Generate a posterior predictive data set}
\usage{
get_predictive_Y(mcmc, nobs, nsets, seed = NULL)
}
\arguments{
\item{\code{mcmcobj}:}{fitted GMC mcmc object from the function \code{GMC.mcmc}}

\item{\code{nobs}:}{number of observations in the posterior predictive data set, default is \code{nrow(data)}}

\item{\code{nsets}:}{number of posterior predictive data sets to create}

\item{\code{seed}:}{seed for replication}
}
\value{
\itemize{
\item \code{Y_pred}: A list of length \code{nets} containing posterior predictive data sets with \code{nobs} observations.
These data sets are constructed \emph{with} the margin adjustment
\item \code{Y_pred_noMA}:  A list of length \code{nets} containing posterior predictive data sets with \code{nobs} observations.
These data sets are constructed \emph{without} the margin adjustment. In this case empirical distribution functions are used
for count variables, and smoothed empirical distributions are used for continuous variables.
\item \code{seed}: seed for replication
}
}
\description{
Utilize posterior samples of Gaussian mixture copula parameters to generate a posterior
predictive data set for missing data inference. Each posterior predictive data set is
constructed with one radomly selected sample from the posterior of GMC parameters.
}
\examples{
mcmc<-GMC.mcmc(Data = X, nsamp = 1000)
pred<- get_predictive_Y(mcmc,
                       nsets = 1,
                       nobs = dim(X)[1],
                       seed = 2)
plot(pred$Y_pred[[1]])

}
