% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Software.R
\name{GMC_Impute}
\alias{GMC_Impute}
\title{Fit Gaussian mixture copula and impute missing data}
\usage{
GMC_Impute(
  Data,
  nImp = 10,
  Impute = T,
  H = 25,
  k.star = NULL,
  nsamp = 10000,
  burn = floor(0.5 * nsamp),
  hyperparams = NULL
)
}
\arguments{
\item{Data}{Data frame that is the input into the Gaussian Mixture Copula. Categorical and  binary variables should be factors}

\item{nImp}{Number of Imputations to Create}

\item{Impute}{logical; Whether or not to create multiple imputations}

\item{H}{Upper bound for the number of components of the mixtures}

\item{k.star}{Dimension of latent factor in mixture}

\item{nsamp}{Number of samples for the MCMC}

\item{burn}{Number of burn-in samples before imputations}

\item{hyperparams}{Hyper parameter specifications}
}
\value{
a list with the following elements:
\itemize{
\item \code{Imputations} nImp complete data sets
\item \code{Deltas} List of length H, each containing an Nsamp - burn x k.star x k.star array of
 posterior samples of cluster specific covariance matrices
\item \code{mus} List of length H, each containing an Nsamp-burn x k.star array of posterior samples of cluster
specific covariance matrices
\item \code{Sigmas} array of dimension (nsamp-burn x p)  containing posterior samples of the error covariance
\item \code{Quantiles} List of length p, each containing an array of dimension (nsamp-burn x length(support)) with posterior samplees
\item \code{z.hat} the estimated latent data (on the transformed scale) at the MLEs
\item \code{residuals} the Dunn-Smyth residuals (randomized)
\item \code{residuals_rep} the Dunn-Smyth residuals (randomized) for 10 replicates
\item \code{logLik} the log-likelihood at the MLEs
\item \code{logLik0} the log-likelihood at the MLEs for the *unrounded* initialization
\item \code{lambda} the Box-Cox nonlinear parameter
\item and other parameters that
(1) track the parameters across EM iterations and
(2) record the model specifications
}
}
\description{
The user specifies this and that
}
\examples{
num = 500
X1<-rnorm(num)
X2<- rpois(num,5*abs(X1))
R = t(sapply(1:num, function(x)rbernoulli(2, p = pnorm(-.5 + .5*abs(X1[x]))))) #change to -.25 for more missing
X<- data.frame(X1,X2,X3)
X_noMis = X
X[which(R[,1] == T),2] = NA
X[which(R[,2] == T),3] = NA
imps<-GMC_Impute(Data = X, nsamp = 1000)
}
